+++ kibble/api/plans_test.go
@@ -15,9 +15,12 @@
 package api
 
 import (
+	"strings"
 	"testing"
 
 	"kibble/models"
+	"kibble/utils"
+
 	"github.com/stretchr/testify/assert"
 )
 
@@ -41,7 +44,7 @@ func TestPlansMapping(t *testing.T) {
 	assert.Equal(t, "", model.Interval)
 }
 
-func TestPlansMappingWithOptionalSvodFields(t *testing.T) {
+func TestRecurringPlanWithExpiryDate(t *testing.T) {
 
 	itemIndex := make(models.ItemIndex)
 
@@ -50,6 +53,8 @@ func TestPlansMappingWithOptionalSvodFields(t *testing.T) {
 	interval := "week"
 	intervalCount := 4
 	trialPeriodDays := 7
+	planType := "recurring"
+	expiryDate := utils.ParseTimeFromString("2021-04-01T03:02:17.000Z")
 
 	apiPlan := PlansV1{
 		Name:            "Bronze Plan",
@@ -58,6 +63,10 @@ func TestPlansMappingWithOptionalSvodFields(t *testing.T) {
 		Interval:        &interval,
 		IntervalCount:   &intervalCount,
 		TrialPeriodDays: &trialPeriodDays,
+		PlanType:        &planType,
+		ExpiryDate:      expiryDate,
+		PortraitImage:   "/portrait_image.png",
+		LandscapeImage:  "/landscape_image.png",
 	}
 
 	model := apiPlan.mapToModel(serviceConfig, itemIndex)
@@ -68,4 +77,36 @@ func TestPlansMappingWithOptionalSvodFields(t *testing.T) {
 	assert.Equal(t, "week", model.Interval)
 	assert.Equal(t, 4, model.IntervalCount)
 	assert.Equal(t, 7, model.TrialPeriodDays)
+	assert.Equal(t, "recurring", model.PlanType)
+	assert.True(t, model.HasExpiryDate())
+	assert.True(t, strings.HasSuffix(model.LandscapeImage, "landscape_image.png"))
+	assert.True(t, strings.HasSuffix(model.PortraitImage, "portrait_image.png"))
+
+}
+
+func TestOneOffPlanWithNoExpiryDate(t *testing.T) {
+	itemIndex := make(models.ItemIndex)
+
+	serviceConfig := commonServiceConfig()
+
+	planType := "one_off"
+
+	apiPlan := PlansV1{
+		Name:        "One Off Plan",
+		Description: "Plan description 123",
+		Status:      "active",
+		PlanType:    &planType,
+	}
+
+	model := apiPlan.mapToModel(serviceConfig, itemIndex)
+	assert.Equal(t, "One Off Plan", model.Name)
+	assert.Equal(t, "one-off-plan", model.NameSlug)
+	assert.Equal(t, "Plan description 123", model.Description)
+	assert.Equal(t, "", model.Interval)
+	assert.Equal(t, 0, model.IntervalCount)
+	assert.Equal(t, 0, model.TrialPeriodDays)
+	assert.Equal(t, "one_off", model.PlanType)
+	assert.False(t, model.HasExpiryDate())
+	assert.Empty(t, model.PortraitImage)
+	assert.Empty(t, model.LandscapeImage)
 }