+++ changelog.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 0.15.22
+Upgrade to go v1.17
+Support for landscape image field in Plans
+
+## 0.15.21
+Add support for Film Award Categories
+
+## 0.15.20
+Add support for new fields in Plans (expiry date and plan type)
+
 ## 0.15.19
 Add support for film info Refs json blob.
 Add new field "Plans" to Page objects, comprising a list of associated Plans which have the current Page set as the showcase